:zap: Ticker performance optimizations
Rewrite of the schema and architecture for tickers. Moves workers
out of the app, and will instead be run by standalone servers.

The schema for tickers has been changed to store an hour of data
in each record rather than the whole set in a subarray. Also,
rather than using an array, an object is used with 60 keys that
represent a minute in the hour, and the minute is also an object
with 60 keys representing a second within a minute. This makes
reading/writing super fast given that we know what key we want.

Also, node-cache is now replaced with redis which is more performant.
More importantly, the redis server is external so that
caching does not use up the app's resources. Also, all the
instances will now have a shared cache instead of each keeping
its own cache.